==> 2_clean_data.ipynb <==
import pandas as pd 

# 1 Limpiando datos 

df_data_historica =  pd.read_csv("fifa_worldcup_historical_data.csv")
df_data_fixture =  pd.read_csv("fifa_worldcup_fixture.csv")

#### 1.1 limpienado df_fixture

df_data_fixture["away"] = df_data_fixture["away"].str.strip()
df_data_fixture["home"] = df_data_fixture["home"].str.strip()

#### 1.2 limpiando df_histotical_data


 index_eliminar = df_data_historica[df_data_historica["home"].str.contains("Sweden") &
                  df_data_historica["away"].str.contains("Austria")].index

df_data_historica.drop(index=index_eliminar, inplace=True)

# df_data_historica[df_data_historica["score"].str.contains("[^\d–]")]
df_data_historica["score"] = df_data_historica["score"].str.replace("[^\d–]", "", regex=True)

df_data_historica["away"] = df_data_historica["away"].str.strip()
df_data_historica["home"] = df_data_historica["home"].str.strip()
#df_data_historica["score"] = df_data_historica["score"].str.strip()

df_data_historica[["HomeGoals", "AwayGoals"]] = df_data_historica["score"].str.split("–", expand=True)

df_data_historica.drop("score", axis=1, inplace=True)
df_data_historica.drop("home_goals", axis=1, inplace=True)
df_data_historica.drop("away_goals", axis=1, inplace=True)

df_data_historica.rename(columns={"home":"HomeTeam",
                                  "away":"AwayTeam",
                                  "year":"Year"}, inplace=True)

df_data_historica.dtypes
df_data_historica = df_data_historica.astype({"HomeGoals":int,
                                              "AwayGoals":int, 
                                              "Year": int})
df_data_historica.dtypes

df_data_historica["TotalGoals"] = df_data_historica["HomeGoals"] + df_data_historica["AwayGoals"]
df_data_historica["TotalGoals"]

df_data_historica.to_csv("clean_fifa_worldcup_historical_data.csv", index=False)
df_data_fixture.to_csv("clean_fifa_worldcup_fixture.csv", index=False)

# Verificar partidos 
years = [1930, 1934, 1938, 1950, 1954, 1958, 1962, 1966, 1970, 1974,
         1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014,
         2018]
for year in years:
    print(year, len(df_data_historica[df_data_historica["Year"] == year]))
